backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
# Use Python 3.11 or 3.12 for pre-built wheels; on 3.14 pip may build from source (needs pg_config).
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.1.0.post1

# Utils
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
# Otp service
redis==5.0.1
